# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.EmbedVideoRequest:
    description: Запрос на создание embed ссылки для видео
    properties:
      url:
        example: https://www.youtube.com/watch?v=abc123
        type: string
    required:
    - url
    type: object
  handlers.EmbedVideoResponse:
    description: Ответ с embed ссылкой для видео
    properties:
      embedUrl:
        example: https://www.youtube.com/embed/abc123
        type: string
      message:
        example: Video embed URL generated
        type: string
      originalUrl:
        example: https://www.youtube.com/watch?v=abc123
        type: string
      type:
        example: youtube
        type: string
    type: object
...
  handlers.InternalErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Internal server error
        type: string
    type: object
  handlers.InternalLoadErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Loading error
        type: string
    type: object
  handlers.InternalProcessingErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Processing error
        type: string
    type: object
...
  handlers.InvalidFileErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid file
        type: string
    type: object
...
  handlers.InvalidParametersErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.InvalidURLErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid URL
        type: string
    type: object
...
  handlers.ProfileResponse:
    description: Ответ с данными профиля пользователя
    properties:
      avatarUrl:
        example: https://example.com/avatars/123.jpg
        type: string
      createdAt:
        example: "2023-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      fullName:
        example: Иван Иванов
        type: string
      id:
        example: 123
        type: integer
      isVerified:
        example: true
        type: boolean
      role:
        example: teacher
        type: string
      specializations:
        example:
        - math
        - physics
        items:
          type: string
        type: array
      updatedAt:
        example: "2023-01-15T10:30:00Z"
        type: string
    type: object
...
  handlers.UpdateProfileRequest:
    description: Запрос на обновление профиля
    properties:
      fullName:
        example: Иван Иванов
        type: string
      specializations:
        example:
        - math
        - physics
        items:
          type: string
        type: array
    type: object
  handlers.UpdateProfileResponse:
    description: Ответ на обновление профиля
    properties:
      data:
        $ref: '#/definitions/handlers.UpdateProfileRequest'
      message:
        example: Profile updated successfully
        type: string
      userID:
        example: 123
        type: integer
    type: object
  handlers.UploadAvatarResponse:
    description: Ответ на загрузку аватара
    properties:
      avatarUrl:
        example: https://example.com/avatars/123.jpg
        type: string
      message:
        example: Avatar uploaded successfully
        type: string
    type: object
  handlers.UploadImageResponse:
    description: Ответ на загрузку изображения
    properties:
      fileName:
        example: image.jpg
        type: string
      height:
        example: 1080
        type: integer
      message:
        example: Image uploaded successfully
        type: string
      size:
        example: 102400
        type: integer
      url:
        example: https://example.com/images/abc123.jpg
        type: string
      width:
        example: 1920
        type: integer
    type: object
  handlers.UploadVideoResponse:
    description: Ответ на загрузку видео
    properties:
      fileName:
        example: video.mp4
        type: string
      message:
        example: Video uploaded successfully
        type: string
      size:
        example: 10485760
        type: integer
      url:
        example: https://example.com/videos/abc123.mp4
        type: string
    type: object
  handlers.UserNotFoundErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: No such user
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для образовательной платформы Пайдея
  title: Paydeya Education Platform API
  version: "1.0"
paths:
...
  /media/embed:
    post:
      consumes:
      - application/json
      description: Создает embed ссылку для видео с YouTube, VK и других сервисов
      parameters:
      - description: URL видео
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.EmbedVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ссылка для вставки создана
          schema:
            $ref: '#/definitions/handlers.EmbedVideoResponse'
        "400":
          description: Неверный URL
          schema:
            $ref: '#/definitions/handlers.InvalidURLErrorResponse'
        "500":
          description: Ошибка обработки
          schema:
            $ref: '#/definitions/handlers.InternalProcessingErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Вставить видео по ссылке
      tags:
      - media
  /media/images:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для использования в материалах
      parameters:
      - description: Файл изображения
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение загружено
          schema:
            $ref: '#/definitions/handlers.UploadImageResponse'
        "400":
          description: Неверный файл
          schema:
            $ref: '#/definitions/handlers.InvalidFileErrorResponse'
        "500":
          description: Ошибка загрузки
          schema:
            $ref: '#/definitions/handlers.InternalLoadErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение
      tags:
      - media
  /media/videos:
    post:
      consumes:
      - multipart/form-data
      description: Загружает видео файл для использования в материалах
      parameters:
      - description: Файл видео
        in: formData
        name: video
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Видео загружено
          schema:
            $ref: '#/definitions/handlers.UploadVideoResponse'
        "400":
          description: Неверный файл
          schema:
            $ref: '#/definitions/handlers.InvalidFileErrorResponse'
        "500":
          description: Ошибка загрузки
          schema:
            $ref: '#/definitions/handlers.InternalLoadErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузить видео
      tags:
      - media
  /profile:
    get:
      consumes:
      - application/json
      description: Возвращает данные профиля текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные профиля
          schema:
            $ref: '#/definitions/handlers.ProfileResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handlers.UserNotFoundErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - profile
    patch:
      consumes:
      - application/json
      description: Обновляет данные профиля пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль обновлен
          schema:
            $ref: '#/definitions/handlers.UpdateProfileResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить профиль
      tags:
      - profile
  /profile/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Загружает аватар пользователя
      parameters:
      - description: Файл аватара (макс. 5MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Аватар загружен
          schema:
            $ref: '#/definitions/handlers.UploadAvatarResponse'
        "400":
          description: Неверный файл
          schema:
            $ref: '#/definitions/handlers.InvalidFileErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузить аватар
      tags:
      - profile
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...